Zmiany typu zmiennych reprezentujących identyfikatory rekordów z long na int.  Zmiany nazewnictwa tych zmiennych - małe zamiast wielkich liter.   Zmiany zakresu szeregu zmiennych z pól klas na zmienne lokalne. Usunięto szereg bezużytecznych kometarzy JavaDoc.

+++ trunk/src/pl/lewica/api/url/ArticleURL.java
@@ -86,7 +86,7 @@ public class ArticleURL implements WebServiceURL {
 	 * @param extension
 	 * @return
 	 */
-	public static String buildURLImage(long id, String extension) {
+	public static String buildURLImage(int id, String extension) {
 		StringBuilder sb	= new StringBuilder(URLDictionary.IMAGE);
 
 		sb.append(id);
@@ -97,7 +97,7 @@ public class ArticleURL implements WebServiceURL {
 	}
 
 
-	public static String buildNameThumbnail(long id, String extension) {
+	public static String buildNameThumbnail(int id, String extension) {
 		// Thumbnails use the th_ prefix
 		StringBuilder sb	= new StringBuilder(PREFIX_THUMBNAIL);
